refactor: extract GrpcConnectionPool implementation into a package (#5623)
