+++ tests/unit/serve/instrumentation/test_networking_histograms.py
@@ -9,7 +9,7 @@ from opentelemetry.sdk.metrics.export import (
     Metric,
 )
 
-from jina.serve.networking import _NetworkingHistograms
+from jina.serve.networking.instrumentation import _NetworkingHistograms
 
 
 @pytest.fixture
