+++ tests/unit/serve/runtimes/gateway/grpc/test_grpc_tls.py
@@ -5,7 +5,7 @@ import pytest
 from docarray import Document
 
 from jina import Client, Flow
-from jina.serve.networking import GrpcConnectionPool
+from jina.serve.networking.utils import send_health_check_sync
 
 
 @pytest.fixture
@@ -52,7 +52,7 @@ def test_grpc_ssl_with_flow_and_client(cert_pem, key_pem, error_log_level):
         with open(cert_pem, 'rb') as f:
             creds = f.read()
 
-        GrpcConnectionPool.send_health_check_sync(
+        send_health_check_sync(
             target=f'localhost:{flow.port}',
             root_certificates=creds,
             tls=True,
