+++ tests/unit/serve/runtimes/head/test_head_runtime.py
@@ -12,7 +12,12 @@ from docarray import Document, DocumentArray
 from jina.clients.request import request_generator
 from jina.enums import PollingType
 from jina.proto import jina_pb2_grpc
-from jina.serve.networking import GrpcConnectionPool
+from jina.serve.networking.utils import (
+    get_available_services,
+    get_default_grpc_options,
+    send_request_sync,
+    send_requests_sync,
+)
 from jina.serve.runtimes.asyncio import AsyncNewLoopRuntime
 from jina.serve.runtimes.head import HeadRuntime
 from jina.types.request import Request
@@ -31,7 +36,7 @@ async def test_regular_data_case(stream):
 
     with grpc.insecure_channel(
         f'{args.host}:{args.port}',
-        options=GrpcConnectionPool.get_default_grpc_options(),
+        options=get_default_grpc_options(),
     ) as channel:
         if stream:
             stub = jina_pb2_grpc.JinaRPCStub(channel)
@@ -66,7 +71,7 @@ def test_message_merging(disable_reduce, stream):
     assert handle_queue.empty()
 
     data_request = _create_test_data_message()
-    result = GrpcConnectionPool.send_requests_sync(
+    result = send_requests_sync(
         [data_request, data_request], f'{args.host}:{args.port}'
     )
     assert result
@@ -87,9 +92,7 @@ def test_uses_before_uses_after():
 
     assert handle_queue.empty()
 
-    result = GrpcConnectionPool.send_request_sync(
-        _create_test_data_message(), f'{args.host}:{args.port}'
-    )
+    result = send_request_sync(_create_test_data_message(), f'{args.host}:{args.port}')
     assert result
     assert _queue_length(handle_queue) == 5  # uses_before + 3 workers + uses_after
     assert len(result.response.docs) == 1
@@ -122,7 +125,7 @@ def test_decompress(monkeypatch):
 
     with grpc.insecure_channel(
         f'{args.host}:{args.port}',
-        options=GrpcConnectionPool.get_default_grpc_options(),
+        options=get_default_grpc_options(),
     ) as channel:
         stub = jina_pb2_grpc.JinaSingleDataRequestRPCStub(channel)
         response, call = stub.process_single_data.with_call(_create_test_data_message())
@@ -157,7 +160,7 @@ def test_dynamic_polling(polling):
 
     with grpc.insecure_channel(
         f'{args.host}:{args.port}',
-        options=GrpcConnectionPool.get_default_grpc_options(),
+        options=get_default_grpc_options(),
     ) as channel:
         stub = jina_pb2_grpc.JinaSingleDataRequestRPCStub(channel)
         response, call = stub.process_single_data.with_call(
@@ -169,7 +172,7 @@ def test_dynamic_polling(polling):
 
     with grpc.insecure_channel(
         f'{args.host}:{args.port}',
-        options=GrpcConnectionPool.get_default_grpc_options(),
+        options=get_default_grpc_options(),
     ) as channel:
         stub = jina_pb2_grpc.JinaSingleDataRequestRPCStub(channel)
         response, call = stub.process_single_data.with_call(
@@ -198,7 +201,7 @@ def test_base_polling(polling):
 
     with grpc.insecure_channel(
         f'{args.host}:{args.port}',
-        options=GrpcConnectionPool.get_default_grpc_options(),
+        options=get_default_grpc_options(),
     ) as channel:
         stub = jina_pb2_grpc.JinaSingleDataRequestRPCStub(channel)
         response, call = stub.process_single_data.with_call(
@@ -210,7 +213,7 @@ def test_base_polling(polling):
 
     with grpc.insecure_channel(
         f'{args.host}:{args.port}',
-        options=GrpcConnectionPool.get_default_grpc_options(),
+        options=get_default_grpc_options(),
     ) as channel:
         stub = jina_pb2_grpc.JinaSingleDataRequestRPCStub(channel)
         response, call = stub.process_single_data.with_call(
@@ -235,7 +238,7 @@ async def test_head_runtime_reflection():
     )
 
     async with grpc.aio.insecure_channel(f'{args.host}:{args.port}') as channel:
-        service_names = await GrpcConnectionPool.get_available_services(channel)
+        service_names = await get_available_services(channel)
 
     assert all(
         service_name in service_names
@@ -258,7 +261,7 @@ def test_timeout_behaviour():
 
     with grpc.insecure_channel(
         f'{args.host}:{args.port}',
-        options=GrpcConnectionPool.get_default_grpc_options(),
+        options=get_default_grpc_options(),
     ) as channel:
         stub = jina_pb2_grpc.JinaSingleDataRequestRPCStub(channel)
         response, call = stub.process_single_data.with_call(_create_test_data_message())
